2213 0209

+++ src/controllers/AgendaController.js
@@ -1,4 +1,5 @@
 const connection= require('../database/connection');
+const sendmail= require('../utils/mailer');
 
 module.exports={
     
@@ -6,7 +7,7 @@ module.exports={
 
         const response= await connection('agenda')
         .join('socios','socios.id','=','agenda.socio_id')
-        .select('agenda.*','socios.cpf','socios.nome');
+        .select('agenda.*','socios.cpf','socios.nome').orderBy('id', 'desc');
         
         return res.status(200).send(response);
     },
@@ -27,21 +28,60 @@ module.exports={
 
     async update(req,res){
         const agenda_id= req.params.id;
-        
+
+        const {email}= await connection('agenda')
+        .join('socios','socios.id','=','agenda.socio_id')
+        .where('agenda.id', agenda_id)
+        .select('socios.email')
+        .first()
+
         await connection('agenda').where('id', agenda_id).update('status', 'confirmado');
+
+        sendmail.agenda_confirmado(email);
 
         return res.status(200).send({message: 'Agenda confirmada com sucesso'});
     },
 
-    async delete(req,res){
+    async reject(req,res){
         const agenda_id= req.params.id;
 
-        await connection('agenda').where('id', agenda_id).delete();
+        const {email}= await connection('agenda')
+        .join('socios','socios.id','=','agenda.socio_id')
+        .where('agenda.id', agenda_id)
+        .select('socios.email')
+        .first()
 
-        return res.status(200).send({message: 'Agenda deletada com sucesso'})
+        await connection('agenda').where('id', agenda_id).delete();
+       
+        sendmail.agenda_recusado(email);
+
+        return res.status(200).send({message: 'Agenda recusada com sucesso'})
     },
 
-    async remove_socio(req,res){
+    async delete(req,res){
+        const socio_id= req.socio_id;
+        const agenda_id= req.params.id;
 
+        const verify= await connection('agenda').where('id', agenda_id).select('socio_id').first();
+
+        if(verify.socio_id!==socio_id){
+            return res.status(401).send({message: 'Esta agenda não lhe pertence'})
+        }
+
+        const data_agenda= await connection('agenda').where('id', agenda_id).select('data').first();
+
+        const now= new Date();
+        const validade= data_agenda.data
+        const parts = validade.split('/');
+        const data = new Date(parts[2], parts[1] - 1, parts[0]);
+
+        if(now>=data){
+            return res.status(401).send({message: 'Você não pode deletar pois este evento já passou'})
+        }
+
+        await connection('agenda').where('id', agenda_id).delete();
+        return res.status(200).send({message: 'Agenda deletada com sucesso'});
+        // const verify= awati
     }
+
 }
